# apps/worker/docs/swagger.yaml
definitions:
  github_com_kunish_wheel_apps_worker_internal_types.BaseUrl:
    properties:
      delay:
        type: integer
      url:
        type: string
    type: object
  github_com_kunish_wheel_apps_worker_internal_types.ChannelCreateRequest:
    properties:
      autoGroup:
        type: integer
      autoSync:
        type: boolean
      baseUrls:
        items:
          $ref: "#/definitions/github_com_kunish_wheel_apps_worker_internal_types.BaseUrl"
        type: array
      customHeader:
        items:
          $ref: "#/definitions/github_com_kunish_wheel_apps_worker_internal_types.CustomHeader"
        type: array
      customModel:
        type: string
      enabled:
        type: boolean
      keys:
        items:
          $ref: "#/definitions/github_com_kunish_wheel_apps_worker_internal_types.ChannelKeyInput"
        type: array
      model:
        items:
          type: string
        type: array
      name:
        type: string
      paramOverride:
        type: string
      type:
        type: integer
    type: object
  github_com_kunish_wheel_apps_worker_internal_types.ChannelEnableRequest:
    properties:
      enabled:
        type: boolean
      id:
        type: integer
    type: object
  github_com_kunish_wheel_apps_worker_internal_types.ChannelKeyInput:
    properties:
      channelKey:
        type: string
      remark:
        type: string
    type: object
  github_com_kunish_wheel_apps_worker_internal_types.CustomHeader:
    properties:
      key:
        type: string
      value:
        type: string
    type: object
...
  internal_handler.fetchModelPreviewRequest:
    properties:
      baseUrl:
        type: string
      key:
        type: string
      type:
        type: integer
    type: object
  internal_handler.fetchModelRequest:
    properties:
      id:
        type: integer
    type: object
...
info:
  contact: {}
  description: LLM API Gateway — aggregate providers, manage load balancing, and track
    usage/costs.
  title: Wheel API
  version: 1.0.0
paths:
...
  /api/v1/channel/create:
    post:
      consumes:
        - application/json
      parameters:
        - description: Channel configuration
          in: body
          name: body
          required: true
          schema:
            $ref: "#/definitions/github_com_kunish_wheel_apps_worker_internal_types.ChannelCreateRequest"
      produces:
        - application/json
      responses:
        "200":
          description: "{success: true, data: Channel}"
          schema:
            type: object
        "400":
          description: "{success: false, error: string}"
          schema:
            type: object
      security:
        - BearerAuth: []
      summary: Create a new channel
      tags:
        - Channels
  /api/v1/channel/delete/{id}:
    delete:
      parameters:
        - description: Channel ID
          in: path
          name: id
          required: true
          type: integer
      produces:
        - application/json
      responses:
        "200":
          description: "{success: true}"
          schema:
            type: object
        "400":
          description: "{success: false, error: string}"
          schema:
            type: object
      security:
        - BearerAuth: []
      summary: Delete a channel
      tags:
        - Channels
  /api/v1/channel/enable:
    post:
      consumes:
        - application/json
      parameters:
        - description: Channel ID and enabled state
          in: body
          name: body
          required: true
          schema:
            $ref: "#/definitions/github_com_kunish_wheel_apps_worker_internal_types.ChannelEnableRequest"
      produces:
        - application/json
      responses:
        "200":
          description: "{success: true}"
          schema:
            type: object
        "400":
          description: "{success: false, error: string}"
          schema:
            type: object
      security:
        - BearerAuth: []
      summary: Enable or disable a channel
      tags:
        - Channels
  /api/v1/channel/fetch-model:
    post:
      consumes:
        - application/json
      parameters:
        - description: Channel ID
          in: body
          name: body
          required: true
          schema:
            $ref: "#/definitions/internal_handler.fetchModelRequest"
      produces:
        - application/json
      responses:
        "200":
          description: "{success: true, data: {models: []string}}"
          schema:
            type: object
        "400":
          description: "{success: false, error: string}"
          schema:
            type: object
      security:
        - BearerAuth: []
      summary: Fetch models from an existing channel
      tags:
        - Channels
  /api/v1/channel/fetch-model-preview:
    post:
      consumes:
        - application/json
      parameters:
        - description: Channel type, base URL, and API key
          in: body
          name: body
          required: true
          schema:
            $ref: "#/definitions/internal_handler.fetchModelPreviewRequest"
      produces:
        - application/json
      responses:
        "200":
          description: "{success: true, data: {models: []string}}"
          schema:
            type: object
        "400":
          description: "{success: false, error: string}"
          schema:
            type: object
      security:
        - BearerAuth: []
      summary: Preview models from channel credentials (without saving)
      tags:
        - Channels
  /api/v1/channel/last-sync-time:
    get:
      produces:
        - application/json
      responses:
        "200":
          description: "{success: true, data: {lastSyncTime: int}}"
          schema:
            type: object
      security:
        - BearerAuth: []
      summary: Get last model sync timestamp
      tags:
        - Channels
  /api/v1/channel/list:
    get:
      produces:
        - application/json
      responses:
        "200":
          description: "{success: true, data: {channels: []Channel}}"
          schema:
            type: object
      security:
        - BearerAuth: []
      summary: List all channels
      tags:
        - Channels
  /api/v1/channel/sync:
    post:
      produces:
        - application/json
      responses:
        "200":
          description: "{success: true, data: SyncResult}"
          schema:
            type: object
      security:
        - BearerAuth: []
      summary: Sync models from all auto-sync channels
      tags:
        - Channels
  /api/v1/channel/update:
    post:
      consumes:
        - application/json
      parameters:
        - description: Partial channel fields to update (id required)
          in: body
          name: body
          required: true
          schema:
            type: object
      produces:
        - application/json
      responses:
        "200":
          description: "{success: true}"
          schema:
            type: object
        "400":
          description: "{success: false, error: string}"
          schema:
            type: object
      security:
        - BearerAuth: []
      summary: Update channel configuration
      tags:
        - Channels
...
  /api/v1/model/channel:
    get:
      produces:
        - application/json
      responses:
        "200":
          description: "{success: true, data: []object}"
          schema:
            type: object
      security:
        - BearerAuth: []
      summary: List models grouped by channel
      tags:
        - Models
...
  /api/v1/stats/channel:
    get:
      produces:
        - application/json
      responses:
        "200":
          description: "{success: true, data: {channels: []ChannelStatsItem}}"
          schema:
            type: object
      security:
        - BearerAuth: []
      summary: Get per-channel statistics
      tags:
        - Stats
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"